Update the toolkit to 1.8.4
Update the toolkit to 1.8.4 to be up to date with the latest 1.8.x
version. Update the Dojox::VERSION constant and the changelog as
well.

Thanks to bamboo, a lot of assets have been reorganized correctly
and some of them have been removed because they were useless (such
as the .uncompressed.js files).

Please also see robin850/dojo-rails@eae0f91

:sparkles: